backend/requirements-local.txt
# For local Qwen-Image-Layered (no API key). Requires CUDA GPU.
# pip install -r requirements-local.txt
diffusers>=0.30.0
transformers>=4.51.3
torch
accelerate
pillow
